create base card item and use to clean up other places of the app
