+++ components/cart/CardList.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
-import { FlatList } from 'react-native';
+import { FlatList, Pressable } from 'react-native';
+import Icon from 'react-native-vector-icons/FontAwesome';
 
-import CardItem from './CardItem';
+import BaseCardItem from './BaseCardItem';
 
 const dummyData = [
     { type: 'visa', cardNumber: '****-****-1234' },
@@ -11,15 +12,18 @@ const dummyData = [
 
 const CardList = () => {
     const [selectedId, setSelectedId] = useState(0);
+
     return (
         <FlatList
             data={dummyData}
             renderItem={({ item, index }) => (
-                <CardItem
-                    {...item}
-                    isSelected={index === selectedId}
-                    onClick={() => setSelectedId(index)}
-                />
+                <Pressable onPress={() => setSelectedId(index)}>
+                    <BaseCardItem {...item}>
+                        {index === selectedId && (
+                            <Icon name="check" color="green" />
+                        )}
+                    </BaseCardItem>
+                </Pressable>
             )}
             keyExtractor={(_, i) => i}
         />
